+++ src/yetibot/core/commands/echo.clj
@@ -5,6 +5,7 @@
 
 (defn echo-cmd
   "echo <text> # Echos back <text>. Useful for piping."
+  {:yb/cat #{:util}}
   [{:keys [args] :as cmd-args}]
   (info "echo cmd args:" cmd-args)
   args)