+++ src/yetibot/core/commands/status.clj
@@ -11,11 +11,13 @@
 
 (defn show-status
   "status # show statuses in the last 8 hours"
+  {:yb/cat #{:util :info}}
   [{:keys [chat-source]}]
   (model/format-sts (model/status-since (pr-str chat-source) (-> 8 hours ago))))
 
 (defn show-status-since
   "status since <n> <minutes|hours|days|weeks|months> ago # show status since a given time"
+  {:yb/cat #{:util :info}}
   [{[_ n unit] :match chat-source :chat-source}]
   (let [unit (plural unit) ; pluralize if singular
         unit-fn (ns-resolve 'clj-time.core (symbol unit))
@@ -26,6 +28,7 @@
 
 (defn set-status
   "status <message> # update your status"
+  {:yb/cat #{:util :info}}
   [{:keys [match chat-source user] :as args}]
   (let [str-chat-source (pr-str chat-source)]
     (info "add status in" str-chat-source ":" match)