+++ src/yetibot/core/commands/uptime.clj
@@ -10,6 +10,7 @@
 
 (defn uptime-cmd
   "uptime # list uptime in milliseconds"
+  {:yb/cat #{:util}}
   [_] (DurationFormatUtils/formatDurationWords (uptime-millis) true true))
 
 (cmd-hook #"uptime"